+++ lib/src/dynamic/bootstrap.dart
@@ -1,7 +1,9 @@
 import 'package:coflui/src/dynamic/builders/action_builders.dart';
 import 'package:coflui/src/dynamic/builders/container_builders.dart';
+import 'package:coflui/src/dynamic/builders/content_builders.dart';
 import 'package:coflui/src/dynamic/builders/display_builders.dart';
 import 'package:coflui/src/dynamic/builders/input_builders.dart';
+import 'package:coflui/src/dynamic/builders/media_builders.dart';
 import 'package:coflui/src/dynamic/models/ui_component.dart';
 import 'package:coflui/src/dynamic/registry/widget_registry.dart';
 
@@ -40,6 +42,15 @@ class DynamicUIBootstrap {
 
       // Action
       UIType.button: ActionBuilders.button,
+
+      // Media
+      UIType.icon: MediaBuilders.icon,
+      UIType.image: MediaBuilders.image,
+      UIType.gradientBar: MediaBuilders.gradientBar,
+
+      // Content
+      UIType.listTile: ContentBuilders.listTile,
+      UIType.detailRow: ContentBuilders.detailRow,
     });
   }
 }